requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Transformers and NLP
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
tokenizers>=0.14.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.7.0

# Configuration Management
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0
pathlib>=1.0.0
typing-extensions>=4.7.0

# Optimizers
lion-pytorch>=0.2.0

# Optional: For better performance
huggingface_hub[hf_xet]>=0.17.0 